+++ api/base/resources/exceptions.py
@@ -1,17 +1,23 @@
 # Authors: Nikolay Melnik <nmelnik@upwork.com>
+from api.base.exceptions import ApiBaseException
 
-__all__ = ('NotFound', 'ValidationError')
+
+__all__ = ('NotFound', 'ValidationError', 'NotImplemented')
 
 
-class NotFound(Exception):
+class NotFound(ApiBaseException):
     pass
 
 
-class ValidationError(Exception):
+class NotImplemented(ApiBaseException):
+    pass
+
+
+class ValidationError(ApiBaseException):
     """
     Base exception class for form validation, that stores
-    validation errors.
+    validation errors and traceback for each error.
     """
-    def __init__(self, *args, **kwargs):
+    def __init__(self, message, chain=None, **kwargs):
         self.errors = kwargs.pop('errors', None)
-        super(ValidationError, self).__init__(*args, **kwargs)
+        super(ValidationError, self).__init__(message, chain)