+++ api/ml_models/tasks/exceptions.py
@@ -1,14 +1,16 @@
 # Authors: Nikolay Melnik <nmelnik@upwork.com>
+from api.base.tasks import TaskException
+
 
 __all__ = ['VisualizationException']
 
 
-class VisualizationException(Exception):
+class VisualizationException(TaskException):
     CLASSIFIER_NOT_SUPPORTED = 1
     ALL_WEIGHT_NOT_FILLED = 2
 
-    def __init__(self, message, error_code=0):
-        super(VisualizationException, self).__init__(message)
+    def __init__(self, message, chain=None, error_code=0):
+        super(VisualizationException, self).__init__(message, chain)
         self.error_code = error_code
 
     def __str__(self):